api_service_demo/requirements.txt
aiohttp==3.1.1
pymongo==3.5.1
Flask==0.12.2
pandas==0.22.0
numpy==1.14.1
PyMySQL==0.8.0
Werkzeug==0.11.11
requests==2.18.4
peewee==2.10.2
redis==2.10.6
Flask_Script==2.0.6
DBUtils==1.2
python_dateutil==2.7.0
PyYAML==3.12
xlrd>=0.9.0
xlsxwriter==1.0.2
gunicorn==19.7.1
